prepare for url validation
