+++ src/api.py
@@ -1,6 +1,7 @@
 from flask import Flask
 from flask import request
 from flask_cors import CORS, cross_origin
+from apiHandler import apiHandler
 
 
 app = Flask(__name__)
@@ -11,3 +12,12 @@ app.config['CORS_HEADERS'] = 'Content-Type'
 @cross_origin()
 def index():
     return "Server Connected"
+
+@app.route('/', methods=["POST"])
+def post_outlet():
+    if request.method == 'POST':
+        get_data=request.args
+        apiHandler.create_url(get_data.to_dict())
+        
+
+    return 'insert successfully'
