+++ src/run.py
@@ -1,5 +1,4 @@
 from api import app
-from logic import *
 
 if __name__ == '__main__':
     app.run(debug=False)
